Rewrote "easy" solver and puzzle solver
The "easy" function has been renamed to "fromText", now able to
take arbitrary text files as an input.

The "solver" function is now recursive. It will try the "easiest"
method first and choose a progressively "harder" solving method
if the current method is unable to uncover any of the puzzle (via
the self.changes flag). If the self.changes flag is changed the
solver reverts to the easy method and the cycle continues until
either the puzzle is solved or it is deemed unsolvable (if one
complete cycle of methods all fail to make changes).

The nakedTwin method is not actually as CPU intensive as first
documented a couple commits ago. The high CPU usage was due
to the solver function continually running whilst the
self.changes flag is true. Looking at the output from the new
recursive solver which shows the methods used, I suspect that
the nakedTwin method is setting the self.changes flag to true
on finding a nakedTwin, regardless of any actual changes to
surrounding flags being made.