NakedTwin no longer incorrectly edits changes flag
Previously the nakedTwin method would naively set the changes flag
to true if it came across a location that was not one of the naked
twins. This lead to a bug that modified the changes flag every time
it had discovered a nakedTwin, even if surrounding locations were
not actually modified at all.

Now, the method checks to see if surrounding locations are proper
supersets of the nakedTwins. This leads to the expected behaviour,
as the surrounding location is not a proper superset of the
nakedTwins, there is nothing that needs to be changed.

Note that the new check replaces the old check as something cannot
be a proper subset of itself, meaning there is no need to check if
the surrounding location is not a nakedTwin location.