Refactored tests
The solver function now has the ability to provide snapshosts
of the solving progress after each method is run.
Unfortunately the string representing the puzzle that is
passed becomes garbled.

created a puzzleSummary function that provides useful info
about a puzzle, such as showing the puzzle before and after
a solving attempt, as well as the amount of values updated
and any updated ghostValues. It also shows the order in which
solving methods were executed (via the history of solver).

fromText has now been reduced to a helper that only passes
select puzzles from a text file to puzzleSummary.
All the analytical parts of fromText have been moved and
enhanced in puzzleSummary, in an attempt to make the system
more modular.