Removed all "breakdown" parameters
As adding the state of the puzzle after every iteration during
the solver method (providing that the breakdown parameter was
set), I decided to remove it entirely.

A step by step snapshot of the puzzle being solved can be generated
alternately through the use of the history, that shows the methods
used to solve a puzzle.