Fixed populateGhosts Method
Previously the populate ghosts method would only attempt to
modify the ghosts of a location that had been passed through
the modified locations variable. This was problematic as there
were of course no ghosts to change, as the naked single method
had already deleted the key from the ghost values dict.

Now we check to see if the group contains any of the locations
that were modified, before looping through all the locations in
the group and removing any ghosts.