NakedN no longer incorrectly sets changes flag
Previously the nakedN method would naively set the changes flag
to true if it came across a location that was not one of the naked
twins. This lead to a bug that modified the changes flag every time
it had discovered a naked set of length n, even if surrounding
locations were not actually modified at all.