Relaxed conditions for nakedN
Previously nakedN only considered a naked group of size n to be
one which had all n ghost values in every location
e.g. (1,2,3), (1,2,3), (1,2,3). This is not always the case, as
(1,2), (2,3) and (1,3) represent a naked triple.

Instead of just checking the first location in combinations, all
locations in the combination are checked to see if they produce
a group of size n.