Made calculateDimensions method for init
The new method means that the gridSize and subGrids parameters
do not need to be manually defined, as they are inferred from the
length of the data string.

For non square grid sizes, factors of the grid sizes are calculated
to set the values of subGridsX and subGridsY. A new flag,
horizontalFormat has been added for non square grid sizes, which
determines if the sub grids are horizontal or not.