Prevented gameLoop (now play) from 100% CPU
As sublimeText does not support text input natively, the game loop
runs continuously waiting for a response that will never come, this
will eventually take up large amounts of memory and CPU.

Fixed by adding an exception counter.

Note that this isn't a real problem in the sense that it should not be
run in an application incapable of text input.