New: getXWingGroups and visualiser
getXWingGroups does what it says, finding all possible X-Wing groups
ready to be analysed by an XWing function. I need to do more research
to see if it is implemented correctly as currently the two locations in
the same row can also be in the same sub grid.

visualiser takes a group of locations as an input as simply prints an
empty puzzle, save for the locations which are marked. This was useful
to see if the groups generated by getXWingGroups looked as expected.

empty groups in intersection groups are no longer pruned, as I found it
easier for the getXWingGroups method to simply look for the row groups
via a known offset, which would be impossible if intersection groups were
not guaranteed to be somewhere.
e.g. in a 9x9 sudoku, there are 9 sub grid groups, followed by 9 row
groups, followed by 9 column groups