Mini overhaul as a result of new intersect groups
xWing method now properly implemented, all four xWing locations cannot
exist in the same subGrid. Also now properly checks to see if the
xWing ghosts occur nowhere else in either rows or columns.

Update Puzzle method now called if changes flag is set.

Get X Neighbours methods are now implemented. Quite expensive as
they loop through all locations. Can be shortening by terminating
when correct number of neighbours found.

Made distinction between generating complete X Groups and pruned
X Groups. More thought needed on how necessary both of them are.

Intersection types are not generated until a puzzle needs to be solved,
this is to save time and memory if solving is not required for a puzzle.
Initialise Intersections is called by every solving method, to ensure
that the subGrid, row and column groups exist before a solving attempt
is made, as they are necessary for ghosts to work. If they are not found,
they are generated and Initialise ghosts is called to setup.
New types of intersections can be passed through as requirements, and as
before if the groups do not exist they will be created.