Added validation for the number of values and grid
Now if the gridSize is a prime number (an invalid configuration), an
exception is raised.

An exception is also raised when an incorrect number of values is
provided. This has been done by looking for the nearest square number,
as the number of values provided must be a square for a puzzle to
be valid.