Added Box Line Reduction method
This new method has been merged with the the pointing N method and
renamed to Intersection Removal. There are now flags to select which
method you want to use.

The reasons for this is because both methods make use of row and
column pointers, and as different size puzzles need to be supported,
it would be difficult to dynamically generate these groups in the
same way as other groups (rows, xWings etc) are.

This will of course impose a reduction in speed as the row groups
have to be generated every time the method is run instead of the
ideal way, generating them once and then pruning after changes.