Fixed generate conjugate pair method
Corrected the conditions for a conjugate pair, which only occur
when a candidate appears twice in the same row, column or subgrid.

Previously a pair of locations were only considered to be conjugate
if they had exactly the same 2 candidates. While this situation is
sufficient it is not necessary, as both locations in the pair need
only share one of its candidates.