Stricter puzzle validation, solve mode flag added
Now a puzzle is marked as invalid when a location is empty, but has
no solving candidates.
To support the new criterion, a solve mode flag has been created,
that is triggered by the initialise intersections method. This is to
try and prevent this criterion being improperly examined, for example
the check may incorrectly show a puzzle to be invalid if it has just
been created as there would be no solving candidates.

The flag should prove useful anyway when a gui is made.