Added helper dictionaries in init
Note: The last commit is dependent on this one, as one new
dictionary in this commit has been referenced in the previous
commit.

Now, there are units that are defined. This will be helpful for
X-Sudokus, as the diagonals can be added to units for easy looping,
without having to worry about what units (intersection types) are
used.

Static Groups have been defined, which have been used in the
validator method.

Neighbour methods have been put in a dictionary to help looping,
now you can loop over all units, and then use the current unit to
find the correct method to find the relevant neighbours.