Fixed chain generation method
Now, each chain only contains a series of consecutive conjugate
pairs with at least one location in common e.g. (1,2,3), where (1,2)
and (3,2) OR (2,3) are conjugate pairs. This is a stronger assertion
than the previous implementation where a location could be added to
a chain solely because it was in a conjugate pair and they could "see"
each other, but had no common locations.