Conjugate Pairs are now added in reverse order too
This change was made to support the chain generation method, as
there were cases in which a chain would be incomplete due to the
fact that only the last link in the chain is compared with other
conjugate pairs.

Instead of duplicating the chain generation method to loop over
conjugate pairs in reverse, chains have been added in reverse. As
this consumes more memory (as every conjugate pair exists in
reverse), it may be less costly to change the chain generator
instead to look at chains in reverse while building them.