Added logging strings to the chain methods
Overall, the logging structure feels quite messy in it's
implementation, especially with more complicated methods such
as the chain methods. The success strings are terribly written
at the moment but that is not a real concern.

Imagining a possible GUI, you can almost pre-empt that this
logging system will have to be re-written. Maybe the "logs"
returned should be very barebones (containing only the
essentials such as locations modified and candidates removed,
no explanation fluff). This is so that a GUI could then do some
fancy highlighting to make it clear what will be changed.

Then again, I'm not sure if the more verbose strings can be
moved from where they are, as of course each solving method has
different logic, so not having a different string makes the whole
feature slightly basic and not very helpful.