Strengthened validation criteria.
As solving candidates are generated (if they do not exist) when
the solving method is called, we should always check to see if
every cell has solving candidates, as if one doesn't the puzzle
is impossible.

This criteria should be used regardless of the solve mode, as
previously a freshly entered but potentially invalid puzzle
would incorrectly show as valid just because it is not in solve
mode.