Fixed initialiseCandidates, getValues
This method previously used getEmptyLocations, which returns the
keys of the candidates dictionary. Of course when the puzzle has
first been made, there are no candidates in the dictionary and so
we have to iterate over the amount of locations and check they are
empty.

Also fixed get Values by removing the unpacking of the list
comprehension, and the union, as the list comprehension is
a flat list. (items in the values dict will always be one
number).