Refactored getAlignment, now named alignment
This method now uses a new alignment methods dictionary instead
of defining the methods in the specific method.

The alignments are now returned sorted, this is to prevent a case
in which two sets of locations potentially have the same
alignment, but the units are in a different order. Previously we
could guarantee what order the units would be search in as we
iterated over a tuple instead of a dictionary.