Added remove solving candidates method
This method takes any number of candidates that may or may not
be in the location wanted. If the candidate is in the location,
it will be removed.

The method returns a list of those candidates that were removed,
so that it is no longer necessary to test each candidate to see
if it was removed or not.