Improved conjugate pair chain building
Now this method is up to par with the chain building method used
for locked pairs.

It will now search for links at either end of the chain, meaning that
longer chains will be found, and that it is no longer necessary
to reserve an initial chain pair.

Subsets of existing chains will now be found and purged, and if a
greater chain is found the existing, smaller chain will be purged.