Simplified Swordfish using neighbours method
There are now no hard to read ways of find elements and candidates
on certain rows and columns.

candidates for rows and columns are now stored in dictionaries
where the keys are the row/column indices, and the values are
sets so that checking the membership of a candidate in a row
or column is simpler.

Now there is no need to check what type of swordfish we have
since we can find neighbours easily.