Fixed dancing links solve method
Previously, the matrix solver would call itself recursively,
this in itself is fine, but the return statement before it
meant that the solver could never get to the backtrack portion,
necessary for harder puzzles.

The removal of this return function however makes it so that the
stack fully unwinds, losing all solutions as there is no completion
check, so one was added.

Finally the order of node neighbour iteration was reversed.